Refactoring v2.0
