+++ eudi/credentials/mdoc/holder.go
@@ -94,49 +94,3 @@ func (h *Holder) SignDeviceAuth(docType string, transcript SessionTranscript) ([
 
 	return cbor.Marshal(msg)
 }
-
-// SelectiveDisclose filters the credential to only include the requested attributes
-// issuerAuth is reused unchanged — the issuer's signature covers all digests regardless
-// of which subset the holder chooses to reveal at any given presentation
-func SelectiveDisclose(mdoc *MDoc, namespace string, reveal []string) (*MDoc, error) {
-	fmt.Println("\n--- HOLDER: Selective disclosure ---")
-
-	revealSet := make(map[string]bool)
-	for _, r := range reveal {
-		revealSet[r] = true
-	}
-
-	allItems := mdoc.IssuerSigned.NameSpaces[namespace]
-	var disclosed []Tag24Item
-
-	for _, tag24item := range allItems {
-		// decode Tag-24 wrapped item to peek at the elementIdentifier
-		var rawTag cbor.RawTag
-		if err := cbor.Unmarshal(tag24item.EncodedItem, &rawTag); err != nil {
-			return nil, fmt.Errorf("unwrap tag24: %w", err)
-		}
-		var innerBytes []byte
-		if err := cbor.Unmarshal(rawTag.Content, &innerBytes); err != nil {
-			return nil, fmt.Errorf("unwrap inner bytes: %w", err)
-		}
-		var item IssuerSignedItem
-		if err := cbor.Unmarshal(innerBytes, &item); err != nil {
-			return nil, fmt.Errorf("decode item: %w", err)
-		}
-
-		if revealSet[item.ElementIdentifier] {
-			fmt.Printf("  Revealing:   %s\n", item.ElementIdentifier)
-			disclosed = append(disclosed, tag24item)
-		} else {
-			fmt.Printf("  Withholding: %s\n", item.ElementIdentifier)
-		}
-	}
-
-	return &MDoc{
-		DocType: mdoc.DocType,
-		IssuerSigned: IssuerSigned{
-			NameSpaces: map[string][]Tag24Item{namespace: disclosed},
-			IssuerAuth: mdoc.IssuerSigned.IssuerAuth, // reused unchanged
-		},
-	}, nil
-}
