+++ eudi/credentials/mdoc/mdoc_test.go
@@ -1,73 +1,14 @@
 package mdoc
 
 import (
-	"bytes"
-	"crypto/x509"
 	"encoding/hex"
-	"strings"
 	"testing"
-	"time"
 
 	"github.com/fxamacker/cbor/v2"
 )
 
 // ============================================================
-// TEST HELPERS
-// ============================================================
-
-// buildHappyPathMDoc runs the full issuer → holder pipeline once and
-// returns everything a verifier needs. Centralized here so every test
-// below starts from the same known-good, real (not hand-crafted) mdoc.
-func buildHappyPathMDoc(t *testing.T) (*Issuer, *Holder, *Verifier, *MDoc, SessionTranscript, []byte, string, string) {
-	t.Helper()
-
-	issuer, err := NewIssuer()
-	if err != nil {
-		t.Fatalf("NewIssuer: %v", err)
-	}
-
-	holder, err := NewHolder()
-	if err != nil {
-		t.Fatalf("NewHolder: %v", err)
-	}
-
-	docType := "eu.europa.ec.av.1"
-	namespace := "eu.europa.ec.av.1"
-
-	claims := map[string]any{
-		"age_over_18": true,
-		"age_over_16": true,
-		"age_over_21": false,
-	}
-
-	mdoc, err := issuer.Issue(docType, namespace, claims, holder.PublicKey())
-	if err != nil {
-		t.Fatalf("Issue: %v", err)
-	}
-
-	presented, err := SelectiveDisclose(mdoc, namespace, []string{"age_over_18"})
-	if err != nil {
-		t.Fatalf("SelectiveDisclose: %v", err)
-	}
-
-	transcript := SessionTranscript{
-		DeviceEngagementBytes: []byte("test-engagement"),
-		EReaderKeyBytes:       []byte("test-reader-key"),
-		Handover:              "test-handover",
-	}
-
-	deviceAuthBytes, err := holder.SignDeviceAuth(docType, transcript)
-	if err != nil {
-		t.Fatalf("SignDeviceAuth: %v", err)
-	}
-
-	verifier := NewVerifier([]*x509.Certificate{issuer.IACACert()})
-
-	return issuer, holder, verifier, presented, transcript, deviceAuthBytes, docType, namespace
-}
-
-// ============================================================
-// TEST 1: FULL HAPPY-PATH FLOW — dumps real mdoc CBOR bytes
+// FULL HAPPY-PATH FLOW — dumps real mdoc CBOR bytes
 // ============================================================
 
 // TestFullIssuanceFlow_ProducesValidMDoc runs issuance → selective
@@ -121,682 +62,5 @@ func TestFullIssuanceFlow_ProducesValidMDoc(t *testing.T) {
 	}
 	if _, present := result.Attributes["age_over_21"]; present {
 		t.Fatalf("age_over_21 should have been withheld, but was disclosed")
-	}
-}
-
-// ============================================================
-// TEST 2: COSE KEY ENCODING — proves the keyasint fix actually works
-// ============================================================
-
-// TestCOSEKeyUsesIntegerMapKeys decodes the real MSO bytes produced by
-// the issuer and checks — at the raw CBOR level — that deviceKeyInfo's
-// map keys are CBOR integers (major type 0/1), not text strings. This
-// is the concrete regression test for the COSEKey struct-tag fix.
-func TestCOSEKeyUsesIntegerMapKeys(t *testing.T) {
-	issuer, holder, _, _, _, _, _, _ := buildHappyPathMDoc(t)
-	_ = issuer
-	_ = holder
-
-	// Re-issue directly so we have the raw MSO payload bytes in hand.
-	docType := "eu.europa.ec.av.1"
-	namespace := "eu.europa.ec.av.1"
-	newHolder, _ := NewHolder()
-	mdoc, err := issuer.Issue(docType, namespace, map[string]any{"age_over_18": true}, newHolder.PublicKey())
-	if err != nil {
-		t.Fatalf("Issue: %v", err)
-	}
-
-	// Decode COSE_Sign1 → MSO payload, then decode the MSO into a generic
-	// map so we can inspect deviceKeyInfo.deviceKey's key types directly,
-	// bypassing our own (possibly-wrong) struct tags.
-	var raw map[string]cbor.RawMessage
-	// issuerAuth is itself a COSE_Sign1; the payload field inside it is
-	// the MSO. Easiest robust check: decode issuerAuth generically.
-	var coseGeneric []any
-	if err := cbor.Unmarshal(mdoc.IssuerSigned.IssuerAuth, &coseGeneric); err != nil {
-		t.Fatalf("decode cose generic: %v", err)
-	}
-	if len(coseGeneric) < 3 {
-		t.Fatalf("expected COSE_Sign1 array with >=3 elements, got %d", len(coseGeneric))
-	}
-	msoPayload, ok := coseGeneric[2].([]byte)
-	if !ok {
-		t.Fatalf("payload element wrong type: %T", coseGeneric[2])
-	}
-
-	if err := cbor.Unmarshal(msoPayload, &raw); err != nil {
-		t.Fatalf("decode mso generic: %v", err)
-	}
-
-	deviceKeyInfoRaw, ok := raw["deviceKeyInfo"]
-	if !ok {
-		t.Fatalf("deviceKeyInfo missing from MSO")
-	}
-	var dkiGeneric map[string]cbor.RawMessage
-	if err := cbor.Unmarshal(deviceKeyInfoRaw, &dkiGeneric); err != nil {
-		t.Fatalf("decode deviceKeyInfo generic: %v", err)
-	}
-	deviceKeyRaw, ok := dkiGeneric["deviceKey"]
-	if !ok {
-		t.Fatalf("deviceKey missing from deviceKeyInfo")
-	}
-
-	// Decode deviceKey as map[any]any to see actual key types.
-	var keyMap map[any]any
-	if err := cbor.Unmarshal(deviceKeyRaw, &keyMap); err != nil {
-		t.Fatalf("decode deviceKey as generic map: %v", err)
-	}
-
-	for k := range keyMap {
-		switch v := k.(type) {
-		case int64:
-			// negative keys decode as int64
-		case uint64:
-			// positive keys decode as uint64
-			_ = v
-		default:
-			t.Fatalf("deviceKey map key %v has type %T, want int64/uint64 (COSEKey struct tags missing ',keyasint')", k, k)
-		}
-	}
-
-	// Sanity: the four expected keys (1, -1, -2, -3) must exist.
-	// Positive key (1) decodes as uint64; negative keys decode as int64.
-	want := []int64{1, -1, -2, -3}
-	for _, w := range want {
-		found := false
-		for k := range keyMap {
-			switch kv := k.(type) {
-			case int64:
-				if kv == w {
-					found = true
-				}
-			case uint64:
-				if w >= 0 && kv == uint64(w) {
-					found = true
-				}
-			}
-			if found {
-				break
-			}
-		}
-		if !found {
-			t.Fatalf("deviceKey missing expected key %d (saw keys: %v)", w, keysOf(keyMap))
-		}
-	}
-}
-
-// keysOf is a small debug helper for readable failure messages.
-func keysOf(m map[any]any) []any {
-	out := make([]any, 0, len(m))
-	for k := range m {
-		out = append(out, k)
-	}
-	return out
-}
-
-// ============================================================
-// TEST 3: DETERMINISTIC DIGEST ORDERING
-// ============================================================
-
-// TestClaimOrderingIsDeterministic issues the same claim set twice and
-// checks that ElementIdentifier→DigestID assignment is identical both
-// times. Regression test for the map-iteration-order fix.
-func TestClaimOrderingIsDeterministic(t *testing.T) {
-	issuer, err := NewIssuer()
-	if err != nil {
-		t.Fatalf("NewIssuer: %v", err)
-	}
-	holder, _ := NewHolder()
-
-	claims := map[string]any{
-		"age_over_18": true,
-		"age_over_16": true,
-		"age_over_21": false,
-		"age_over_65": false,
-	}
-
-	extractOrder := func(mdoc *MDoc, namespace string) []string {
-		items := mdoc.IssuerSigned.NameSpaces[namespace]
-		order := make([]string, len(items))
-		for _, tag24item := range items {
-			var rawTag cbor.RawTag
-			if err := cbor.Unmarshal(tag24item.EncodedItem, &rawTag); err != nil {
-				t.Fatalf("unwrap tag24: %v", err)
-			}
-			var innerBytes []byte
-			if err := cbor.Unmarshal(rawTag.Content, &innerBytes); err != nil {
-				t.Fatalf("unwrap inner: %v", err)
-			}
-			var item IssuerSignedItem
-			if err := cbor.Unmarshal(innerBytes, &item); err != nil {
-				t.Fatalf("decode item: %v", err)
-			}
-			order[item.DigestID] = item.ElementIdentifier
-		}
-		return order
-	}
-
-	namespace := "eu.europa.ec.av.1"
-
-	mdoc1, err := issuer.Issue("eu.europa.ec.av.1", namespace, claims, holder.PublicKey())
-	if err != nil {
-		t.Fatalf("Issue #1: %v", err)
-	}
-	mdoc2, err := issuer.Issue("eu.europa.ec.av.1", namespace, claims, holder.PublicKey())
-	if err != nil {
-		t.Fatalf("Issue #2: %v", err)
-	}
-
-	order1 := extractOrder(mdoc1, namespace)
-	order2 := extractOrder(mdoc2, namespace)
-
-	if len(order1) != len(order2) {
-		t.Fatalf("order length mismatch: %v vs %v", order1, order2)
-	}
-	for i := range order1 {
-		if order1[i] != order2[i] {
-			t.Fatalf("digestID→identifier order not deterministic: run1=%v run2=%v", order1, order2)
-		}
-	}
-
-	// Expect alphabetical: age_over_16, age_over_18, age_over_21, age_over_65
-	want := []string{"age_over_16", "age_over_18", "age_over_21", "age_over_65"}
-	for i, w := range want {
-		if order1[i] != w {
-			t.Fatalf("expected sorted order %v, got %v", want, order1)
-		}
-	}
-}
-
-// ============================================================
-// TEST 4: NEGATIVE CASES — these must all FAIL verification
-// ============================================================
-
-func TestUntrustedRootIsRejected(t *testing.T) {
-	_, _, verifier, _, _, _, docType, namespace := buildHappyPathMDoc(t)
-
-	attackerIssuer, _ := NewIssuer()
-	attackerHolder, _ := NewHolder()
-	attackerMDoc, err := attackerIssuer.Issue(docType, namespace,
-		map[string]any{"age_over_18": true}, attackerHolder.PublicKey())
-	if err != nil {
-		t.Fatalf("attacker Issue: %v", err)
-	}
-	attackerPresented, err := SelectiveDisclose(attackerMDoc, namespace, []string{"age_over_18"})
-	if err != nil {
-		t.Fatalf("attacker SelectiveDisclose: %v", err)
-	}
-
-	result := verifier.Verify(attackerPresented, namespace)
-	if result.Valid {
-		t.Fatalf("expected attacker mdoc (untrusted root) to be rejected, but it was accepted")
-	}
-}
-
-func TestTamperedDigestIsRejected(t *testing.T) {
-	_, _, verifier, presented, _, _, _, namespace := buildHappyPathMDoc(t)
-
-	tamperedItem := IssuerSignedItem{
-		DigestID:          0,
-		Random:            []byte("attacker-does-not-know-real-salt"),
-		ElementIdentifier: "age_over_18",
-		ElementValue:      false, // flipped from true
-	}
-	tamperedWrapped, err := tag24Wrap(tamperedItem)
-	if err != nil {
-		t.Fatalf("tag24Wrap: %v", err)
-	}
-	tamperedMDoc := &MDoc{
-		DocType: presented.DocType,
-		IssuerSigned: IssuerSigned{
-			NameSpaces: map[string][]Tag24Item{namespace: {{EncodedItem: tamperedWrapped}}},
-			IssuerAuth: presented.IssuerSigned.IssuerAuth,
-		},
-	}
-
-	result := verifier.Verify(tamperedMDoc, namespace)
-	if result.Valid {
-		t.Fatalf("expected tampered claim to be rejected, but it was accepted")
-	}
-}
-
-func TestDeviceAuthWrongSignerIsRejected(t *testing.T) {
-	_, _, verifier, presented, transcript, _, docType, namespace := buildHappyPathMDoc(t)
-
-	// A different holder (i.e. a different device) signs deviceAuth for
-	// the SAME session transcript, but their key isn't the one embedded
-	// in this mdoc's deviceKeyInfo. Simulates a cloned/copied mdoc.
-	otherHolder, _ := NewHolder()
-	wrongDeviceAuth, err := otherHolder.SignDeviceAuth(docType, transcript)
-	if err != nil {
-		t.Fatalf("SignDeviceAuth: %v", err)
-	}
-
-	result := verifier.VerifyWithDeviceAuth(presented, namespace, docType, transcript, wrongDeviceAuth)
-	if result.DeviceAuthValid {
-		t.Fatalf("expected deviceAuth signed by wrong device key to be rejected, but it was accepted")
-	}
-	// VerifyWithDeviceAuth intentionally marks the overall result invalid
-	// when device binding fails — a presentation without a valid device
-	// signature is not a valid presentation, even if issuerAuth/digests
-	// check out on their own. Confirm the underlying issuerAuth checks
-	// were in fact what ran (via the error message), rather than some
-	// earlier unrelated failure.
-	if result.Valid {
-		t.Fatalf("expected overall result to be invalid when deviceAuth fails, but Valid was true")
-	}
-	if result.Error == "" {
-		t.Fatalf("expected a descriptive error when deviceAuth fails, got empty string")
-	}
-}
-
-func TestDeviceAuthWrongSessionIsRejected(t *testing.T) {
-	_, holder, verifier, presented, transcript, _, docType, namespace := buildHappyPathMDoc(t)
-
-	// Same (correct) device key, but signs over a DIFFERENT session
-	// transcript than the one the verifier actually used. Simulates a
-	// replayed deviceAuth from an earlier/different session.
-	otherTranscript := SessionTranscript{
-		DeviceEngagementBytes: []byte("different-engagement"),
-		EReaderKeyBytes:       []byte("different-reader-key"),
-		Handover:              "different-handover",
-	}
-	replayedDeviceAuth, err := holder.SignDeviceAuth(docType, otherTranscript)
-	if err != nil {
-		t.Fatalf("SignDeviceAuth: %v", err)
-	}
-
-	// Verifier checks against the ORIGINAL transcript it actually issued.
-	result := verifier.VerifyWithDeviceAuth(presented, namespace, docType, transcript, replayedDeviceAuth)
-	if result.DeviceAuthValid {
-		t.Fatalf("expected deviceAuth bound to a different session to be rejected, but it was accepted")
-	}
-}
-
-func TestUnknownDigestIDIsRejected(t *testing.T) {
-	_, _, verifier, presented, _, _, _, namespace := buildHappyPathMDoc(t)
-
-	// Same value/salt shape as a real item, but digestID 999 doesn't
-	// exist in the MSO's valueDigests map at all.
-	bogusItem := IssuerSignedItem{
-		DigestID:          999,
-		Random:            bytes.Repeat([]byte{0x01}, 16),
-		ElementIdentifier: "age_over_18",
-		ElementValue:      true,
-	}
-	bogusWrapped, err := tag24Wrap(bogusItem)
-	if err != nil {
-		t.Fatalf("tag24Wrap: %v", err)
-	}
-	bogusMDoc := &MDoc{
-		DocType: presented.DocType,
-		IssuerSigned: IssuerSigned{
-			NameSpaces: map[string][]Tag24Item{namespace: {{EncodedItem: bogusWrapped}}},
-			IssuerAuth: presented.IssuerSigned.IssuerAuth,
-		},
-	}
-
-	result := verifier.Verify(bogusMDoc, namespace)
-	if result.Valid {
-		t.Fatalf("expected unknown digestID to be rejected, but it was accepted")
-	}
-}
-
-// TestFreshCertsVerifyUnderCurrentTime is a sanity check that chain
-// verification's CurrentTime handling isn't broken by an off-by-one in
-// NotBefore/NotAfter arithmetic — i.e. that a cert issued "now" with a
-// 1-year validity window actually verifies "now".
-func TestFreshCertsVerifyUnderCurrentTime(t *testing.T) {
-	issuer, err := NewIssuer()
-	if err != nil {
-		t.Fatalf("NewIssuer: %v", err)
-	}
-	verifier := NewVerifier([]*x509.Certificate{issuer.IACACert()})
-
-	holder, _ := NewHolder()
-	mdoc, err := issuer.Issue("eu.europa.ec.av.1", "eu.europa.ec.av.1",
-		map[string]any{"age_over_18": true}, holder.PublicKey())
-	if err != nil {
-		t.Fatalf("Issue: %v", err)
-	}
-	presented, err := SelectiveDisclose(mdoc, "eu.europa.ec.av.1", []string{"age_over_18"})
-	if err != nil {
-		t.Fatalf("SelectiveDisclose: %v", err)
-	}
-
-	result := verifier.Verify(presented, "eu.europa.ec.av.1")
-	if !result.Valid {
-		t.Fatalf("expected freshly issued DS cert to be valid under current time, got: %s", result.Error)
-	}
-}
-
-// TestExpiredDSCertIsRejected uses NewVerifierWithClock to pin the
-// verifier's notion of "now" to a point roughly 400 days in the future —
-// past the DS cert's 1-year (365 day) validity window — and checks the
-// chain is correctly rejected as expired. This exercises the actual
-// expiry-rejection path, unlike the sanity check above.
-func TestExpiredDSCertIsRejected(t *testing.T) {
-	issuer, err := NewIssuer()
-	if err != nil {
-		t.Fatalf("NewIssuer: %v", err)
-	}
-
-	holder, _ := NewHolder()
-	mdoc, err := issuer.Issue("eu.europa.ec.av.1", "eu.europa.ec.av.1",
-		map[string]any{"age_over_18": true}, holder.PublicKey())
-	if err != nil {
-		t.Fatalf("Issue: %v", err)
-	}
-	presented, err := SelectiveDisclose(mdoc, "eu.europa.ec.av.1", []string{"age_over_18"})
-	if err != nil {
-		t.Fatalf("SelectiveDisclose: %v", err)
-	}
-
-	futureClock := time.Now().Add(400 * 24 * time.Hour) // past the DS cert's 365-day NotAfter
-	verifier := NewVerifierWithClock([]*x509.Certificate{issuer.IACACert()}, futureClock)
-
-	result := verifier.Verify(presented, "eu.europa.ec.av.1")
-	if result.Valid {
-		t.Fatalf("expected DS cert to be rejected as expired when checked 400 days in the future, but it was accepted")
-	}
-	t.Logf("correctly rejected expired chain: %s", result.Error)
-}
-
-// TestExpiredMSOValidityIsRejected uses the verifier's clock to check the
-// MSO's OWN validUntil (90 days from issuance per Issuer.Issue), separately
-// from the X.509 DS cert's 365-day expiry. A clock ~100 days out is past the
-// MSO's window but still well within the DS cert's — this specifically
-// exercises the mso.ValidityInfo check, not the certificate chain check.
-func TestExpiredMSOValidityIsRejected(t *testing.T) {
-	issuer, err := NewIssuer()
-	if err != nil {
-		t.Fatalf("NewIssuer: %v", err)
-	}
-	holder, _ := NewHolder()
-	mdoc, err := issuer.Issue("eu.europa.ec.av.1", "eu.europa.ec.av.1",
-		map[string]any{"age_over_18": true}, holder.PublicKey())
-	if err != nil {
-		t.Fatalf("Issue: %v", err)
-	}
-	presented, err := SelectiveDisclose(mdoc, "eu.europa.ec.av.1", []string{"age_over_18"})
-	if err != nil {
-		t.Fatalf("SelectiveDisclose: %v", err)
-	}
-
-	futureClock := time.Now().Add(100 * 24 * time.Hour) // past MSO's 90-day validUntil, well within DS cert's 365-day window
-	verifier := NewVerifierWithClock([]*x509.Certificate{issuer.IACACert()}, futureClock)
-
-	result := verifier.Verify(presented, "eu.europa.ec.av.1")
-	if result.Valid {
-		t.Fatalf("expected credential to be rejected as expired per MSO validityInfo, but it was accepted")
-	}
-	if !strings.HasPrefix(result.Error, "credential expired") {
-		t.Fatalf("expected the MSO validityInfo check specifically to fail, got a different error: %s", result.Error)
-	}
-	t.Logf("correctly rejected on MSO validityInfo: %s", result.Error)
-}
-
-// TestNotYetValidMSOIsRejected mirrors the above but checks the ValidFrom
-// side. The clock is pinned to 2 minutes before "now" — AFTER the certs'
-// backdated NotBefore (-5 minutes, see Issuer cert templates) so the X.509
-// chain check passes, but BEFORE the MSO's validFrom (~"now", set in
-// Issue()) so only the MSO validityInfo check can fail. Asserts on the
-// specific "credential not yet valid" prefix so this can't pass for the
-// wrong reason (e.g. coincidentally matching the cert-chain error text).
-func TestNotYetValidMSOIsRejected(t *testing.T) {
-	issuer, err := NewIssuer()
-	if err != nil {
-		t.Fatalf("NewIssuer: %v", err)
-	}
-	holder, _ := NewHolder()
-	mdoc, err := issuer.Issue("eu.europa.ec.av.1", "eu.europa.ec.av.1",
-		map[string]any{"age_over_18": true}, holder.PublicKey())
-	if err != nil {
-		t.Fatalf("Issue: %v", err)
-	}
-	presented, err := SelectiveDisclose(mdoc, "eu.europa.ec.av.1", []string{"age_over_18"})
-	if err != nil {
-		t.Fatalf("SelectiveDisclose: %v", err)
-	}
-
-	pastClock := time.Now().Add(-2 * time.Minute) // after cert NotBefore (-5min), before MSO validFrom (~now)
-	verifier := NewVerifierWithClock([]*x509.Certificate{issuer.IACACert()}, pastClock)
-
-	result := verifier.Verify(presented, "eu.europa.ec.av.1")
-	if result.Valid {
-		t.Fatalf("expected credential to be rejected as not-yet-valid per MSO validityInfo, but it was accepted")
-	}
-	if !strings.HasPrefix(result.Error, "credential not yet valid") {
-		t.Fatalf("expected the MSO validityInfo check specifically to fail, got a different error (chain check probably fired first): %s", result.Error)
-	}
-	t.Logf("correctly rejected on MSO validityInfo: %s", result.Error)
-}
-
-// TestNotYetValidCertIsRejected pins the verifier's clock to a point
-// BEFORE the certs' NotBefore (i.e. before issuance), which should also
-// fail chain verification.
-func TestNotYetValidCertIsRejected(t *testing.T) {
-	issuer, err := NewIssuer()
-	if err != nil {
-		t.Fatalf("NewIssuer: %v", err)
-	}
-
-	holder, _ := NewHolder()
-	mdoc, err := issuer.Issue("eu.europa.ec.av.1", "eu.europa.ec.av.1",
-		map[string]any{"age_over_18": true}, holder.PublicKey())
-	if err != nil {
-		t.Fatalf("Issue: %v", err)
-	}
-	presented, err := SelectiveDisclose(mdoc, "eu.europa.ec.av.1", []string{"age_over_18"})
-	if err != nil {
-		t.Fatalf("SelectiveDisclose: %v", err)
-	}
-
-	pastClock := time.Now().Add(-24 * time.Hour) // before NotBefore
-	verifier := NewVerifierWithClock([]*x509.Certificate{issuer.IACACert()}, pastClock)
-
-	result := verifier.Verify(presented, "eu.europa.ec.av.1")
-	if result.Valid {
-		t.Fatalf("expected DS cert to be rejected as not-yet-valid when checked before issuance, but it was accepted")
-	}
-	t.Logf("correctly rejected not-yet-valid chain: %s", result.Error)
-}
-
-// ============================================================
-// AV PROFILE CLAIM VALIDATION (Annex A §4.1.2)
-// ============================================================
-
-func TestIssueRejectsDisallowedAttribute(t *testing.T) {
-	issuer, err := NewIssuer()
-	if err != nil {
-		t.Fatalf("NewIssuer: %v", err)
-	}
-	holder, _ := NewHolder()
-
-	_, err = issuer.Issue("eu.europa.ec.av.1", "eu.europa.ec.av.1",
-		map[string]any{
-			"age_over_18": true,
-			"family_name": "Smith", // not permitted per Annex A §4.1.2
-		}, holder.PublicKey())
-
-	if err == nil {
-		t.Fatalf("expected Issue to reject family_name, but it succeeded")
-	}
-	if !strings.Contains(err.Error(), "family_name") {
-		t.Fatalf("expected error to name the offending attribute, got: %v", err)
-	}
-	t.Logf("correctly rejected disallowed attribute: %v", err)
-}
-
-func TestIssueRejectsNonBooleanValue(t *testing.T) {
-	issuer, err := NewIssuer()
-	if err != nil {
-		t.Fatalf("NewIssuer: %v", err)
-	}
-	holder, _ := NewHolder()
-
-	_, err = issuer.Issue("eu.europa.ec.av.1", "eu.europa.ec.av.1",
-		map[string]any{
-			"age_over_18": true,
-			"age_over_21": "true", // string, not bool — must be rejected
-		}, holder.PublicKey())
-
-	if err == nil {
-		t.Fatalf("expected Issue to reject a non-bool value, but it succeeded")
-	}
-	if !strings.Contains(err.Error(), "age_over_21") {
-		t.Fatalf("expected error to name the offending attribute, got: %v", err)
-	}
-	t.Logf("correctly rejected non-bool value: %v", err)
-}
-
-func TestIssueRejectsMissingMandatoryAgeOver18(t *testing.T) {
-	issuer, err := NewIssuer()
-	if err != nil {
-		t.Fatalf("NewIssuer: %v", err)
-	}
-	holder, _ := NewHolder()
-
-	_, err = issuer.Issue("eu.europa.ec.av.1", "eu.europa.ec.av.1",
-		map[string]any{
-			"age_over_21": true, // age_over_18 missing — mandatory per profile
-		}, holder.PublicKey())
-
-	if err == nil {
-		t.Fatalf("expected Issue to reject a claim set missing age_over_18, but it succeeded")
-	}
-	if !strings.Contains(err.Error(), "age_over_18") {
-		t.Fatalf("expected error to mention the missing mandatory attribute, got: %v", err)
-	}
-	t.Logf("correctly rejected missing mandatory attribute: %v", err)
-}
-
-func TestIssueAcceptsValidAgeOverNNVariants(t *testing.T) {
-	issuer, err := NewIssuer()
-	if err != nil {
-		t.Fatalf("NewIssuer: %v", err)
-	}
-	holder, _ := NewHolder()
-
-	// age_over_18 plus a few different NN variants, all bool — should issue cleanly.
-	_, err = issuer.Issue("eu.europa.ec.av.1", "eu.europa.ec.av.1",
-		map[string]any{
-			"age_over_18": true,
-			"age_over_16": true,
-			"age_over_65": false,
-		}, holder.PublicKey())
-
-	if err != nil {
-		t.Fatalf("expected valid age_over_NN variants to be accepted, got: %v", err)
-	}
-}
-
-// ============================================================
-// SPEC ALIGNMENT (AV Blueprint Annex A §A.11 worked example)
-// ============================================================
-
-// TestValidityInfoUsesRFC3339Tag decodes the real MSO bytes generically
-// and confirms signed/validFrom/validUntil are encoded as CBOR tag 0
-// (RFC3339 date-time string), matching the exact encoding shown in the
-// spec's own worked example (`"signed": 0("2025-06-20T08:45:29Z")`) —
-// not a bare Unix epoch integer, which is what this program produced
-// before this fix and would not match a spec-conformant decoder's
-// expectations for a real interop scenario.
-func TestValidityInfoUsesRFC3339Tag(t *testing.T) {
-	issuer, err := NewIssuer()
-	if err != nil {
-		t.Fatalf("NewIssuer: %v", err)
-	}
-	holder, _ := NewHolder()
-	mdoc, err := issuer.Issue("eu.europa.ec.av.1", "eu.europa.ec.av.1",
-		map[string]any{"age_over_18": true}, holder.PublicKey())
-	if err != nil {
-		t.Fatalf("Issue: %v", err)
-	}
-
-	// Decode issuerAuth -> MSO payload bytes generically, without going
-	// through our own MSO struct (which would just decode successfully
-	// either way) — inspect the raw CBOR tag on the wire instead.
-	var coseGeneric []any
-	if err := cbor.Unmarshal(mdoc.IssuerSigned.IssuerAuth, &coseGeneric); err != nil {
-		t.Fatalf("decode cose generic: %v", err)
-	}
-	msoPayload, ok := coseGeneric[2].([]byte)
-	if !ok {
-		t.Fatalf("payload element wrong type: %T", coseGeneric[2])
-	}
-
-	var raw map[string]cbor.RawMessage
-	if err := cbor.Unmarshal(msoPayload, &raw); err != nil {
-		t.Fatalf("decode mso generic: %v", err)
-	}
-	validityRaw, ok := raw["validityInfo"]
-	if !ok {
-		t.Fatalf("validityInfo missing from MSO")
-	}
-
-	var viGeneric map[string]cbor.RawMessage
-	if err := cbor.Unmarshal(validityRaw, &viGeneric); err != nil {
-		t.Fatalf("decode validityInfo generic: %v", err)
-	}
-
-	for _, field := range []string{"signed", "validFrom", "validUntil"} {
-		fieldRaw, ok := viGeneric[field]
-		if !ok {
-			t.Fatalf("validityInfo.%s missing", field)
-		}
-		// A CBOR tag-0 value's first byte is 0xc0 (major type 6, tag 0).
-		if len(fieldRaw) == 0 || fieldRaw[0] != 0xc0 {
-			t.Fatalf("validityInfo.%s is not tag-0 encoded (first byte: %#x) — expected RFC3339 string per spec example", field, fieldRaw[0])
-		}
-	}
-}
-
-// TestDeviceAuthPayloadIsDetached confirms the transmitted deviceAuth
-// COSE_Sign1 has payload = null (detached content), matching the exact
-// wire format shown in the spec's deviceSignature example:
-//
-//	"deviceSignature": [ h'a10126', {}, null, h'...' ]
-//
-// Detached payload means the actual DeviceAuthentication bytes are
-// signed but never sent — the verifier reconstructs them itself and
-// supplies them before verifying (see VerifyWithDeviceAuth).
-func TestDeviceAuthPayloadIsDetached(t *testing.T) {
-	holder, err := NewHolder()
-	if err != nil {
-		t.Fatalf("NewHolder: %v", err)
-	}
-	transcript := SessionTranscript{
-		DeviceEngagementBytes: []byte("test-engagement"),
-		EReaderKeyBytes:       []byte("test-reader-key"),
-		Handover:              "test-handover",
-	}
-	deviceAuthBytes, err := holder.SignDeviceAuth("eu.europa.ec.av.1", transcript)
-	if err != nil {
-		t.Fatalf("SignDeviceAuth: %v", err)
-	}
-
-	var arr []any
-	if err := cbor.Unmarshal(deviceAuthBytes, &arr); err != nil {
-		t.Fatalf("decode deviceAuth generic: %v", err)
-	}
-	if len(arr) != 4 {
-		t.Fatalf("expected 4-element COSE_Sign1 array, got %d elements", len(arr))
-	}
-	if arr[2] != nil {
-		t.Fatalf("expected detached payload (nil/null), got %T: %v", arr[2], arr[2])
-	}
-}
-
-// TestDeviceAuthStillVerifiesWithDetachedPayload confirms that detaching
-// the payload doesn't break verification — VerifyWithDeviceAuth must
-// still succeed by reconstructing the payload itself before checking the
-// signature, exactly as a real verifier would.
-func TestDeviceAuthStillVerifiesWithDetachedPayload(t *testing.T) {
-	_, _, verifier, presented, transcript, deviceAuthBytes, docType, namespace := buildHappyPathMDoc(t)
-
-	result := verifier.VerifyWithDeviceAuth(presented, namespace, docType, transcript, deviceAuthBytes)
-	if !result.DeviceAuthValid {
-		t.Fatalf("expected deviceAuth to verify successfully despite detached payload, got: %s", result.Error)
 	}
 }
